backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.1
Flask-SQLAlchemy==3.1.1
nltk==3.9.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
PyJWT==2.8.0
cryptography==42.0.8
requests==2.32.3
gunicorn==23.0.0
waitress==3.0.2
